+++ Sora/Sora.cs
@@ -65,6 +65,7 @@ public class Sora : IDisposable
     public class Config
     {
         public string SignalingUrl = "";
+        public string[] SignalingUrlCandidate = new string[0];
         public string ChannelId = "";
         public string ClientId = "";
         public string Metadata = "";
@@ -190,7 +191,17 @@ public class Sora : IDisposable
 
         var cc = new SoraConf.Internal.ConnectConfig();
         cc.unity_version = UnityEngine.Application.unityVersion;
-        cc.signaling_url = config.SignalingUrl;
+        if (config.SignalingUrl.Trim().Length != 0)
+        {
+            cc.signaling_url.Add(config.SignalingUrl.Trim());
+        }
+        foreach (var url in config.SignalingUrlCandidate)
+        {
+            if (url.Trim().Length != 0)
+            {
+                cc.signaling_url.Add(url.Trim());
+            }
+        }
         cc.channel_id = config.ChannelId;
         cc.client_id = config.ClientId;
         cc.metadata = config.Metadata;
@@ -281,16 +292,16 @@ public class Sora : IDisposable
         commandBuffer.Clear();
     }
 
-    private delegate void TrackCallbackDelegate(uint track_id, IntPtr userdata);
+    private delegate void TrackCallbackDelegate(uint trackId, string connectionId, IntPtr userdata);
 
     [AOT.MonoPInvokeCallback(typeof(TrackCallbackDelegate))]
-    static private void TrackCallback(uint trackId, IntPtr userdata)
+    static private void TrackCallback(uint trackId, string connectionId, IntPtr userdata)
     {
-        var callback = GCHandle.FromIntPtr(userdata).Target as Action<uint>;
-        callback(trackId);
+        var callback = GCHandle.FromIntPtr(userdata).Target as Action<uint, string>;
+        callback(trackId, connectionId);
     }
 
-    public Action<uint> OnAddTrack
+    public Action<uint, string> OnAddTrack
     {
         set
         {
@@ -303,7 +314,7 @@ public class Sora : IDisposable
             sora_set_on_add_track(p, TrackCallback, GCHandle.ToIntPtr(onAddTrackHandle));
         }
     }
-    public Action<uint> OnRemoveTrack
+    public Action<uint, string> OnRemoveTrack
     {
         set
         {