+++ src/unity.cpp
@@ -44,8 +44,8 @@ void* sora_create() {
 
 void sora_set_on_add_track(void* p, track_cb_t on_add_track, void* userdata) {
   auto sora = (sora::Sora*)p;
-  sora->SetOnAddTrack([on_add_track, userdata](ptrid_t track_id) {
-    on_add_track(track_id, userdata);
+  sora->SetOnAddTrack([on_add_track, userdata](ptrid_t track_id, std::string connection_id) {
+    on_add_track(track_id, connection_id.c_str(), userdata);
   });
 }
 
@@ -53,8 +53,8 @@ void sora_set_on_remove_track(void* p,
                               track_cb_t on_remove_track,
                               void* userdata) {
   auto sora = (sora::Sora*)p;
-  sora->SetOnRemoveTrack([on_remove_track, userdata](ptrid_t track_id) {
-    on_remove_track(track_id, userdata);
+  sora->SetOnRemoveTrack([on_remove_track, userdata](ptrid_t track_id, std::string connection_id) {
+    on_remove_track(track_id, connection_id.c_str(), userdata);
   });
 }